refactor(ssr): move `ClassList` to own file
